Improved metric repository for per-db methods. Closes #921
